refactor(datatable): add UploadedAssetDataTableType.php
It is the overview of the uploaded asset for publishers
